requirements-dev.txt
appdirs==1.4.3
attrs==19.3.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
entrypoints==0.3
fake-bpy-module-2.82==20200218
flake8==3.8.3
flake8-black==0.1.1
flake8-bugbear==20.1.4
flake8-polyfill==1.0.2
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
parameterized==0.7.4
pathspec==0.8.0
pep8-naming==0.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
regex==2020.4.4
requests==2.23.0
toml==0.10.1
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.2
unittest-xml-reporting==3.0.2
urllib3==1.25.9
zipp==3.1.0
